Split data: nav items
